Changed prev course to only accept classes with course and subject entered, fixed bug where you could add multiple of the same person to one session, and added creation of a CSV file for yourself to send to others

+++ app/src/main/java/com/example/birdsofafeather/FakedMessageListener.java
@@ -101,11 +101,13 @@ public class FakedMessageListener extends MessageListener {
                 }
 
                 List<String> peopleInSession = db.sessionsDao().get(sessionID).peopleIDs;
-                peopleInSession.add(personId);
-                Session updatedSession = new Session(sessionID, db.sessionsDao().get(sessionID).sessionName);
-                updatedSession.peopleIDs = peopleInSession;
-                db.sessionsDao().delete(db.sessionsDao().get(sessionID));
-                db.sessionsDao().insert(updatedSession);
+                if(!peopleInSession.contains(personId)) {
+                    peopleInSession.add(personId);
+                    Session updatedSession = new Session(sessionID, db.sessionsDao().get(sessionID).sessionName);
+                    updatedSession.peopleIDs = peopleInSession;
+                    db.sessionsDao().delete(db.sessionsDao().get(sessionID));
+                    db.sessionsDao().insert(updatedSession);
+                }
 
 
                 if(!db.personWithCoursesDao().exists(personId)){
